pyproject.toml
[project]
name = "django-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-filter>=24.3",
    "django-structlog>=9.0.0",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.3.1",
    "drf-spectacular>=0.28.0",
    "pycryptodome>=3.21.0",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "ruff>=0.8.2",
]
